Catch Jena Exception. Offering Upload -> Description Upload
