+++ src/main/webapp/WEB-INF/views/base.jsp
@@ -108,9 +108,9 @@
                 <span class="fa fa-archive fa-fw"></span>
                 <span class="text-plain">My offerings</span>
               </div>
-              <a class="list-group-item" href="${ pageContext.request.contextPath }/offerings/register">
+              <a class="list-group-item" href="${ pageContext.request.contextPath }/descriptions/register">
                 <span class="fa fa-upload fa-fw"></span>
-                <span class="text-plain">Upload new offerings</span>
+                <span class="text-plain">Upload new descriptions</span>
               </a>
             </div>
         <div class="list-group">
